[CCUBE-1522-2][JY] Add individual element-specific yup schemas
