+++ src/components/common/pills/pills.tsx
@@ -3,8 +3,8 @@ import { Controller, useFormContext } from "react-hook-form";
 import { TogglePair } from "src/components/common/toggle-pair/toggle-pair";
 import { MandatoryFieldBox } from "src/components/element-editor/basic-details/common";
 import {
-    ITextarea,
-    ITextareaFieldAttributes,
+    ITextareaAttributes,
+    TTextarea,
     useBuilder,
 } from "src/context-providers";
 import { BottomPositionIcon } from "../icons/bottom-position-icon";
@@ -16,10 +16,10 @@ export const Pills = () => {
     // CONST, STATE, REF
     // =========================================================================
     const { control, unregister, setValue, watch, resetField } =
-        useFormContext<ITextareaFieldAttributes>();
+        useFormContext<ITextareaAttributes>();
 
     const { focusedElement } = useBuilder();
-    const element: ITextarea = focusedElement?.element;
+    const element = focusedElement?.element as TTextarea;
     const watchPills = watch("pills");
 
     // =========================================================================
