+++ src/context-providers/builder/element.types.ts
@@ -61,7 +61,7 @@ export interface IBaseAttributes {
 }
 
 export interface IValidation {
-    validationType?: string;
+    validationType?: EValidationType;
     validationRule?: string;
     validationErrorMessage?: string;
 }
@@ -105,47 +105,67 @@ export interface IBaseFieldAttributes extends IBaseAttributes {
     required: boolean;
     requiredErrorMsg?: string | undefined;
     columns: IColumns;
-    placeholder?: string;
     description?: string;
     preselectedValue?: string;
     conditionalRendering?: IConditionalRendering[];
     prefill?: IPrefillAttributes[];
 }
 
-export interface ITextareaFieldAttributes
-    extends IBaseTextBasedFieldAttributes {
-    resizableInput?: boolean;
-    pills?: boolean;
+// =============================================================================
+// ELEMENT SPECIFIC ATTRIBUTES
+// =============================================================================
+
+export interface IEmailFieldAttributes extends IBaseFieldAttributes {
+    placeholder?: string;
+    validation?: IValidation[];
+}
+
+export interface ITextFieldAttributes extends IBaseFieldAttributes {
+    placeholder?: string;
+    validation?: IValidation[];
+}
+
+export interface ITextareaAttributes extends IBaseFieldAttributes {
+    placeholder?: string;
+    validation?: IValidation[];
+    resizableInput: boolean;
+    pills: boolean;
     pillItems?: IPillItemAttributes[];
     pillPosition?: "top" | "bottom" | null;
 }
 
-export interface IBaseTextBasedFieldAttributes extends IBaseFieldAttributes {
+export interface IContactFieldAttributes extends IBaseFieldAttributes {
+    placeholder?: string;
     validation?: IValidation[];
 }
 
-export interface IBaseOptionGroupBasedFieldAttributes
-    extends IBaseFieldAttributes {
+export interface INumericFieldAttributes extends IBaseFieldAttributes {
+    placeholder?: string;
+    validation?: IValidation[];
+}
+
+export interface IDropdownAttributes extends IBaseFieldAttributes {
+    placeholder?: string;
     dropdownItems?: IDropdownItemAttributes[];
 }
 
 // =============================================================================
 // ELEMENT TYPES (Same order as EElementType)
 // =============================================================================
-export type IContactField = IBaseTextBasedFieldAttributes;
-export type IEmailField = IBaseTextBasedFieldAttributes;
-export type INumericField = IBaseTextBasedFieldAttributes;
-export type ITextField = IBaseTextBasedFieldAttributes;
-export type ITextarea = ITextareaFieldAttributes;
-export type IDropdown = IBaseOptionGroupBasedFieldAttributes;
+export type TContactField = IContactFieldAttributes;
+export type TEmailField = IEmailFieldAttributes;
+export type TNumericField = INumericFieldAttributes;
+export type TTextField = ITextFieldAttributes;
+export type TTextarea = ITextareaAttributes;
+export type TDropdown = IDropdownAttributes;
 
 export type TTextBasedElement =
-    | IEmailField
-    | ITextField
-    | ITextarea
-    | INumericField
-    | IContactField;
+    | TEmailField
+    | TTextField
+    | TTextarea
+    | TNumericField
+    | TContactField;
 
-export type TOptionGroupBasedElement = IDropdown;
+export type TOptionGroupBasedElement = TDropdown;
 
 export type TElement = TTextBasedElement | TOptionGroupBasedElement;
