+++ src/translator/helper.ts
@@ -9,12 +9,12 @@ import {
     TElementMap,
 } from "src/context-providers/builder";
 import { ELEMENT_CONDITION_TYPES, SCHEMA_CONDITION_TYPES } from "src/data";
-import { TextBasedField } from "./text-based-field";
-import { IPrefillConfig } from "./types";
 import {
     PREFILL_ACTIONID_REGEX,
     PREFILL_PATH_REGEX,
-} from "src/yup-schemas/base-yup-schema-helper";
+} from "src/yup-schemas/helper";
+import { TextBasedField } from "./text-based-field";
+import { IPrefillConfig } from "./types";
 
 export const createPrefillObject = (elements: TElementMap) => {
     const prefill = Object.values(elements).reduce((acc, element) => {
