+++ src/yup-schemas/elements/dropdown/dropdown-yup-schema.ts
@@ -1,6 +1,6 @@
 import * as yup from "yup";
 
-export const OPTION_GROUP_BASED_SCHEMA = yup.object().shape({
+export const DROPDOWN_YUP_SCHEMA = yup.object().shape({
     placeholder: yup.string().optional(),
     dropdownItems: yup
         .array()
@@ -54,9 +54,4 @@ export const OPTION_GROUP_BASED_SCHEMA = yup.object().shape({
                 throw new yup.ValidationError(errors);
             }
         ),
-    preselectedValue: yup.string().optional(),
 });
-
-export type TOptionGroupBasedSchema = yup.InferType<
-    typeof OPTION_GROUP_BASED_SCHEMA
->;
