+++ src/yup-schemas/elements/dropdown/index.ts
@@ -1,0 +1,1 @@
+export * from "./dropdown-yup-schema";
