+++ src/yup-schemas/elements/email/index.ts
@@ -1,0 +1,1 @@
+export * from "./email-yup-schema";
