+++ src/yup-schemas/elements/index.ts
@@ -1,4 +1,3 @@
-export * from "./contact";
 export * from "./dropdown";
 export * from "./email";
 export * from "./long-text";
