+++ src/yup-schemas/elements/long-text/index.ts
@@ -1,0 +1,1 @@
+export * from "./long-text-yup-schema";
