+++ src/yup-schemas/elements/numeric/index.ts
@@ -1,0 +1,1 @@
+export * from "./numeric-yup-schema";
