+++ src/yup-schemas/yup-schema-builder.ts
@@ -1,9 +1,14 @@
 import { useBuilder } from "src/context-providers";
 import * as yup from "yup";
 import { EElementType } from "../context-providers/builder/element.types";
-import { BaseYupSchemaHelper } from "./base-yup-schema-helper";
-import { OPTION_GROUP_BASED_SCHEMA } from "./option-group-based-fields";
-import { TEXT_AREA_SCHEMA, TEXT_BASED_SCHEMA } from "./text-based-fields";
+import {
+    DROPDOWN_YUP_SCHEMA,
+    EMAIL_YUP_SCHEMA,
+    LONG_TEXT_YUP_SCHEMA,
+    NUMERIC_YUP_SCHEMA,
+    TEXT_YUP_SCHEMA,
+} from "./elements";
+import { generateBaseYupSchema } from "./helper";
 import { TYupSchema } from "./types";
 
 export namespace YupSchemaBuilder {
@@ -11,29 +16,34 @@ export namespace YupSchemaBuilder {
         try {
             const { elements, focusedElement } = useBuilder();
 
-            const baseSchema = BaseYupSchemaHelper.getBaseYupSchema(
-                elements,
-                focusedElement
-            );
+            const baseSchema = generateBaseYupSchema(elements, focusedElement);
 
             switch (type) {
                 case EElementType.EMAIL:
+                    return yup
+                        .object()
+                        .concat(baseSchema)
+                        .concat(EMAIL_YUP_SCHEMA);
                 case EElementType.TEXT:
+                    return yup
+                        .object()
+                        .concat(baseSchema)
+                        .concat(TEXT_YUP_SCHEMA);
                 case EElementType.NUMERIC:
                     return yup
                         .object()
                         .concat(baseSchema)
-                        .concat(TEXT_BASED_SCHEMA(type));
+                        .concat(NUMERIC_YUP_SCHEMA);
                 case EElementType.TEXTAREA:
                     return yup
                         .object()
                         .concat(baseSchema)
-                        .concat(TEXT_AREA_SCHEMA());
+                        .concat(LONG_TEXT_YUP_SCHEMA);
                 case EElementType.DROPDOWN:
                     return yup
                         .object()
                         .concat(baseSchema)
-                        .concat(OPTION_GROUP_BASED_SCHEMA);
+                        .concat(DROPDOWN_YUP_SCHEMA);
             }
         } catch (error) {
             console.error("Error in schema helper:", error);
